Initial port to droidmedia.
We can start the camera pipeline but there is a crash when we stop because we destroy
our DroidMediaCamera "object" then try to free a buffer.
We also don't render anything to the screen yet.
In addition to that: no fences, image capture or video recording have been tested. We also don't return
preview buffers back for an unknown reason