droidenc: Switch H264 to AVC
This also changes how codec_data is being constructed
GstDroidCodec has a pointer to a function that constructs the codec data GstBuffer.
MPEG4 is more or less a standard memcpy()
H264 needs to parse the stagefright codec data in order to reconstruct the
AVC decoder configuration struct.

This change breaks the H264 encoder as we still do not output the correct NALs for AVC