droiddec: switch H264 AVC and allow manipulating encoded data before decoding it.
It's better to use AVC since we can get a separate codec_data which we can pass
to droidcodec. This allows us to recreate the droidcodec any time we need.

If we stick to byte-stream and recreate the codec then we will not get the SPS/PPS
info to pass as part of the first frame which will break the decoder.

Android still expects NALus to begin with 00 00 00 01 so we process
the encoded data and add this prefix to make it happy.

This all comes at the cost of an extra memcpy but it should not be a big issue.