droiddec: rework _flush
We cannot flush the frames being decoded from the decoder. There is simply no way
to do that. The best we can do is to clear the queue of frames to be encoded.
The problem now is if we get flushed we will still decode the previous queued frames
and push them later on when they get decoded.
This will lead to frames being repeated if the flush happens in the beginning
or inaccurate seeking.
We will just mark the decoder as "dirty" so the next handle_frame can recreate it