droiddec: destroy the codec after we create it.
If we have a very short video file then the decoder will start, decode and get done
before we even get any caps negotiation error from downstream. We cannot simply flush it
because the codec thread in mediacodec is not running. We can still detect that condition
but it's better if we destroy and recreate the codec as I highly doubt the it can be
reused anymore for decoding.