droiddec: get all the info we need from DroidMediaBuffer before we destroy it.
The buffer is tied to the GstMemory which we destroy then blindly try to get
buffer width, height, crop and time stamp which gives us invalid values.

The solution is to get all the info we need before we drop our memory ref
and while at it, assign NULL to the buffer.