droiddec: drop our frame ref before we finish it.
If we push the frame while we still having a ref then the buffer being pushed through the pad will have
refcnt 2 instead of 1. If it reaches the identity element in the video branch like that it will end up
copying the buffer which is not a good idea because our memory cannot be copied.

This used to produce green flashes during playback when we were setting the memory size to 0 but
now after commit dad20 it will cause a segfault