droidcamsrc: rework caps negotiation
We delegate caps negotiation to a mode class which takes care of that and knows
when to restart the preview.

We always negotiate vfsrc but negotiate imgsrc only in image mode and vidsrc only in video mode

This should fix the race condition upon startup when we try to change preview size without
stopping the preview.

It also allows supporting devices which do not provide separate preview modes for video
recording and instead rely on preview size (Support for those is not yet implemented)