droiddec: don't provide incorrect buffer dimensions if we have a crop rectangle
Some decoders provide report a change in resolution and use a higher resolution in order
to accommodate HW constrains but use a crop rect to indicate the correct buffer resolution.

In this case we were setting the wrong caps and buffer meta width and height.

In order to fix that we have to delay configuring the decoder state until we have the first frame
available because that has the info we need.