droiddec: rework I420 output.
Seems the correct way to get I420 from Android decoders is to request raw YUV data and convert
it afterwards to I420 with droidmediaconvert/I420 color conversion.

Since we need to determine what the downstream element expects when we create the codec, part of
our caps negotiation had to be moved back to set_format() which fills in a GstVideoInfo structure
which then can be used later when we are creating the actual codec.