[droidcamsrc] don't use strtof for converting strings to floats. Fixes JB#29553
strtof is locale dependent. If the decimal point for a locale is a comma ","
or anything other than a dot "." then we will fail to convert any strings which we
get from camera HAL to floats because they use dots as the decimal separator.

This causes us to fail to get the ev-compensation-step. If we do not have the
checks introduced in the previous commit then we end up dividing a float value by zero,
getting a garbage value which we happily pass to camera HAL which will blow up.