[droidvdec] don't trust output width and height reported by OMX. Contributes to JB#29726
Some decoders mis-implement (or do not obey) the specifications. The output format will be
identical to the input format but in reality it is not.

An example is a MPEG4 decoder that tells us that the output is 1280x720 but the underlying
gralloc buffers were allocated for 1280x736 buffers.
This will lead to green stride being rendered to the screen. The best thing we can do here
is not to trust OMX and use the width and height from the gralloc buffer to configure our state.