[droidvdec] Make sure not to start the source pad task if we have a downwards state transition. Fixes JB#27706
This sequence can actually happen:
set_format()
change_state()
handle_frame()

change_state() will stop the task which will be started afterwards by handle_frame()
later on GStreaner will deadlock because the task is actually running.

We introduce a new boolean called running that is protected by the stream lock and will
be set to true in _start() and to false in stop_loop().
handle_frame() will then check for that and stop processing the buffers if it is false.
This guarantees that the task will never start after GStreamer has stopped it and not expecting
it to run.