[droidcamsrc] don't start the device if it was not running while negotiating mode caps
If we are in paused state and our caps or mode get changed, we will start the device
unconditionally after negotiating the caps which leads to a device running in paused state.

When changing the mode:
the new mode gets activated and then a call to gst_droidcamsrc_apply_mode_settings()
happens. If we fail that call then we can end up in an error situation but the device
is running.

In the above mentioned situations, a state change from PAUSED to READY will happen
but it will never stop the device because it is only started when we go to PLAYING and
we will end up with an activated buffer pool that will happily crash when we unref it
while destroying the device.