[droidcamsrc] apply camera parameters if we are not starting the camera. Contributes to JB#26923
If we are negotiating our pads and the device is not running then this will happen:
1) negotiate
2) store negotiation parameters
3) reload camera parameters to recalculate max-zoom

This will cause us to what we have negotiated and stored.

A solution is to apply the parameters if we are not going to start the device so
we do not lose them.

This was causing the captured image to not be what the application really requested
because we discard what has been negotiated and go with HAL default.