droidcamsrc: disable flash completely when we switch off torch
The previous logic restored the stored flash setting but that does not switch off
the torch light. We force it to off here to make sure it is really off.
We just do not change the stored value so it gets set properly when we switch back to image mode.