[droidvdec] Don't rely on stride to decide on external buffer usage. Fixes JB#30627
It could be that the Y component stride is sane but the U and V strides are not.
This can happen if the video width is not dividable by eight.
An example is a video whose width is 460. The U stride should be 232 because
it should be dividable by 4 too (And should be at least half the width).

To avoid doing the exact calculations as video-info.c, we will simply
rely on buffer size. This should be safe since any change in strides will
lead to an increase in them and an corresponding increase in the size.