[droidvdec] Don't try to stop the loop if we are already finishing. Contributes to JB#27706
This is another deadlock while stopping the decoder case.

If we are already finishing the codec then don't stop the task when we transition to PAUSED state.
The previous finish is still ongoing and will never get completed because we are acquiring the
stream lock.
If we attempt to finish the codec while we are already finishing it