droidvenc: destroy the codec after we drain it.
It's unknown whether the encoder is still usable or not after it gets drained. In addition to that
the encoder might retain some buffers (Nexus 5 does that) and will never return them back until
it is destroyed. Those buffers will cause droidcamsrc to loop waiting for them to be freed
before it can complete video recording stop sequence.

If we do not wait for them then stagefright will assert complaining that the codec did not
return all the buffers.