droidcodec: Add quirks and a use-codec-supplied-height quirk
Quirks are codec specific and are added to the gstdroidcodecs.conf file.
decoders go under decoder-quirks and encoders go under encoder-quirks

Each key of the above mentioned groups corresponds to the Android mime type of the codec.

The use-codec-supplied-height quirk is used to instruct the decoder to use
the codec reported buffer height when calculating the size of the I420 buffer
during color space conversion.

Some Android adaptations need this.