[droidcamsrc] make deferred property set return true
When a photograghic property is set, if the possible values has not been
retrieved yet, droidcamsrcphotography will save the value and return
FALSE. However, as the value is to be applied later, returning FALSE
is misleading, as that means the action has no effect.

This patch make these deferred property set return TRUE, indicating that
the value is set successfully. Also, the logging message is updated to
indicate the defferred nature of it.