[droidcamsrc] implement camerabin2's preview-image interface
This patch implement camerabin2's preview-image interface,
which make the element post a preview-image bus message when
post-previews property is enabled.

As Android doesn't seem to have the concept of preview-image (there are
raw & postview callbacks but they seem not to be called), I replicate
what Qt does with it's native Android plugin, which is storing the
latest viewfinder frame and using that as the preview. This require
changing how the preview callback flags is calculated as otherwise
preview callback won't be called. The last preview buffer is kept
within the structure of dev but use the seperated lock and condition
variable, as the buffer is updated in the callback called from Android
side.

Camerabin2's preview-image interface also require that the element has
the preview-caps property and must send the preview-image according to
this caps. To fulfill that, a camerabinpipeline is constructed inside
the gstdroidcamsrc element. This pipeline will take care of all
conversion required to make the preview image comply to the provided
caps and will also actually posting it to the bus.

When the image capture is requested, the post-preview property is
copied to image capturing structure so that the change after the
capture won't affect this capture. When the shutter callback is
called and if the preview image is requested, the kept viewfinder
buffer is sent to the preview pipeline for processing & posting. For
video recording, preview is posted when the recording is requested.

This can remove the requirement to patch QtMultimedia not to set the
post-preview flag, and improve the correctness of the element outside
of Mer and Sailfish OS.