[droidcamsrc] don't fail device switch while running
This matches the behavior of gstahc, a camerabin-compatible element for
Android's Camera API, which is in gst-plugins-bad upstream. The set
device will take effect next time the camera is opened.

A useless call to gst_droidcamsrc_photography_init() is also removed.
That function does nothing if src->photo is already initialized.