[droidcamsrc] fix pool failure in raw preview mode
Commit b0a25efff149 ("[gst-droid] Use the buffer pool proposed by
droideglsink in camera and video decoder.") make it possible that
dev->pool might be NULL. However, some code isn't updated for that,
causing failure when vfsrc is using raw video.

This patch adds appropriate checks and/or move code around to make
sure that dev->pool will be accessed only if it's not NULL. This fixes
raw preview mode.