[gst-droid] Make the fps-range the same like gst_droidcamsrc_params_choose_video_framerate to set narrow fps range by the maximal value.
The Qualcomm camera HAL uses the max value of preview-fps-range, but MTK camera HAL uses the min value of preview-fps-range.
The setting maximal values of min and max is doing the work of the viewfinder faster on MTK devices and still fast work of viewfinder on Qualcomm devices.
No any regression.