[gst-droid] Circumvent critical error 'index < array->len'
Under certain conditions the following critical error is triggered:

GStreamer-CRITICAL **: 21:38:48.274: gst_query_set_nth_allocation_pool:
  assertion 'index < array->len' failed

Situations in which this happens include:
1. Destroying a video before it's been played.
2. Streaming a video out two outputs.

It appears to be caused by a call to gst_droidvdec_decide_allocation()
where the query has no allocated pools. In this case a pool is created
as a fallback, but because the query is empty, when an attempt is made
to add the fallback pool, there's no entry for it to be inserted into.
This triggers the critical error when the array length is zero.

This change adds the pool to the query explicitly if these specific
conditions are satisified, thereby ensuring there's an entry in the
query for the pool allocation to be added to.